+++ pypsh/__init__.py
@@ -8,5 +8,5 @@ given regex.
 """
 
 __author__ = 'Mathias Fußenegger'
-__version__ = '0.6.0'
+__version__ = '0.7.0'
 __license__ = 'MIT'